Show git hash in web UI version; fix %zd; killall on install
- main.c: use DRIVER_VERSION (includes git hash) in papplMainloop
  and footer so web UI shows e.g. "0.1.0-abc1234" not "0.1.0"
- driver.c: fix %zd → %ld cast in no-response warning (papplLogPrinter
  doesn't handle %zd on this platform)
- Makefile: add killall before systemctl stop to catch manually-started
  instances that systemd doesn't own

+++ src/driver.c
@@ -601,7 +601,7 @@ static bool hl5170dn_status(pappl_printer_t *printer)
 
     if (bytes <= 0) {
         papplLogPrinter(printer, PAPPL_LOGLEVEL_WARN,
-            "status: no response from printer (bytes=%zd)", bytes);
+            "status: no response from printer (bytes=%ld)", (long)bytes);
         return true;
     }
     buf[bytes] = '\0';
